+++ stores/_test-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @internal/lint
 
+## 0.0.112
+
+### Patch Changes
+
+- Updated dependencies [[`55adddf`](https://github.com/mastra-ai/mastra/commit/55adddfda2a170b00c112bf37d677e8ce5b65d5a)]:
+  - @mastra/core@1.52.1
+
 ## 0.0.111
 
 ### Patch Changes